+++ src/src_platform_jfx/fmg/jfx/mosaic/MosaicCanvasView.java
@@ -79,7 +79,7 @@ public class MosaicCanvasView extends MosaicJfxView<Canvas, Image, MosaicDrawMod
     /** переустанавливаю заного размер мины/флага для мозаики */
     protected void changeSizeImagesMineFlag() {
         MosaicDrawModel<Image> model = getModel();
-        int sq = (int)model.getCellAttr().getSq(model.getPenBorder().getWidth());
+        double sq = model.getCellAttr().getSq(model.getPenBorder().getWidth());
         if (sq <= 0) {
             System.err.println("Error: too thick pen! There is no area for displaying the flag/mine image...");
             sq = 3; // ат балды...
